+++ src/plugins/github/pluginName.js
@@ -1,0 +1,3 @@
+// @flow
+
+export const PLUGIN_NAME = "sourcecred/github-beta";
